Replaced APIResponse class with a struct.
git-svn-id: http://anidbmini.googlecode.com/svn/trunk@22 f167d671-08da-b78c-e0d6-9b4f167803b8

+++ AniDBmini/AniDBmini/Converters/FileSizeConverter.cs
@@ -4,7 +4,7 @@ using System.Windows.Data;
 
 namespace AniDBmini
 {
-	[ValueConversion(typeof(long), typeof(string))]
+    [ValueConversion(typeof(double), typeof(string))]
 	class FileSizeConverter : IValueConverter
 	{
 		public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
